ask upsell dialog
