+++ packages/web/src/app/(app)/@sidebar/components/upgradeBadge.tsx
