+++ packages/web/src/app/(app)/settings/license/page.tsx
@@ -5,7 +5,6 @@ import { Button } from "@/components/ui/button";
 import { ExternalLink } from "lucide-react";
 import { redirect } from "next/navigation";
 import { ActivationCodeCard } from "./activationCodeCard";
-import { CheckoutSuccessModal } from "./checkoutSuccessModal";
 import { OnlineLicenseCard } from "./onlineLicenseCard";
 import { OfflineLicenseCard } from "./offlineLicenseCard";
 import { RecentInvoicesCard } from "./recentInvoicesCard";
@@ -17,7 +16,7 @@ type LicensePageProps = {
     searchParams?: Promise<Record<string, string | string[] | undefined>>;
 } & Record<string, unknown>;
 
-export default authenticatedPage<LicensePageProps>(async ({ prisma, org, user }, props) => {
+export default authenticatedPage<LicensePageProps>(async ({ prisma, org }, props) => {
     const searchParams = await props.searchParams;
     if (searchParams?.refresh === 'true') {
         // Side-trips to the Stripe portal (add PM, manage sub) include
@@ -47,8 +46,6 @@ export default authenticatedPage<LicensePageProps>(async ({ prisma, org, user },
     const invoicesResult = license ? await getAllInvoices() : null;
     const invoices = invoicesResult && !isServiceError(invoicesResult) ? invoicesResult : [];
 
-    const showCheckoutSuccess = searchParams?.checkout === 'success' && !license;
-
     return (
         <div className="flex flex-col gap-6">
             <div>
@@ -74,7 +71,6 @@ export default authenticatedPage<LicensePageProps>(async ({ prisma, org, user },
             {license && <OnlineLicenseCard license={license} />}
             {license && <RecentInvoicesCard invoices={invoices} />}
             {!offlineLicense && !license && <ActivationCodeCard />}
-            {showCheckoutSuccess && <CheckoutSuccessModal userEmail={user.email} />}
         </div>
     );
 }, {
