+++ packages/web/src/ee/features/lighthouse/checkoutReturnHandler.tsx
@@ -1,0 +1,23 @@
+"use client";
+
+import { useSearchParams } from "next/navigation";
+import { LicenseActivactionDialog } from "./licenseActivactionDialog";
+
+interface PostCheckoutHandlerProps {
+    userEmail?: string | null;
+}
+
+// Layout-mounted handler that drives the post-Stripe activation flow regardless
+// of which page the user lands on after checkout. Detects `session_id` in the
+// URL (set by Stripe's substitution of `{CHECKOUT_SESSION_ID}` in successUrl),
+// and renders the claim + activate modal when present.
+export function CheckoutReturnHandler({ userEmail }: PostCheckoutHandlerProps) {
+    const searchParams = useSearchParams();
+    const sessionId = searchParams.get("session_id");
+
+    if (!sessionId) {
+        return null;
+    }
+
+    return <LicenseActivactionDialog userEmail={userEmail} />;
+}
